[K32W0] Restructure RAM storage
RAM storage class was moved to k32w0 folder.

RamStorageKey files were removed and implementation was moved
inside RamStorage.
