--- src/platform/nxp/k32w/common/RamStorageKey.h
@@ -1,42 +1,0 @@
-/*
- *
- *    Copyright (c) 2022 Project CHIP Authors
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-#pragma once
-
-#include <platform/nxp/k32w/common/RamStorage.h>
-
-namespace chip {
-namespace DeviceLayer {
-namespace Internal {
-
-class RamStorageKey
-{
-public:
-    RamStorageKey(RamStorage * storage, uint8_t keyId, uint8_t pdmId);
-    CHIP_ERROR Read(uint8_t * buf, uint16_t & sizeToRead) const;
-    CHIP_ERROR Write(const uint8_t * buf, uint16_t length);
-    CHIP_ERROR Delete();
-
-    static uint16_t GetPdmId(uint8_t keyId, uint8_t pdmId) { return static_cast<uint16_t>(keyId) << 8 | pdmId; }
-
-private:
-    RamStorage * mStorage;
-    uint16_t mId;
-};
-
-} // namespace Internal
-} // namespace DeviceLayer
-} // namespace chip
