+++ src/platform/nxp/k32w/k32w0/BUILD.gn
@@ -27,9 +27,6 @@ if (chip_enable_openthread) {
 static_library("k32w0") {
   sources = [
     "../../../SingletonConfigurationManager.cpp",
-    "../common/RamStorage.cpp",
-    "../common/RamStorageKey.cpp",
-    "../common/RamStorageKey.h",
     "BLEManagerImpl.cpp",
     "BLEManagerImpl.h",
     "../common/BLEManagerCommon.cpp",
@@ -53,6 +50,7 @@ static_library("k32w0") {
     "NFCManagerImpl.h",
     "PlatformManagerImpl.cpp",
     "PlatformManagerImpl.h",
+    "RamStorage.cpp",
     "SystemTimeSupport.cpp",
     "ble_function_mux.c",
   ]
@@ -62,8 +60,6 @@ static_library("k32w0") {
     "${chip_root}/src/credentials/examples/DeviceAttestationCredsExample.h",
     "${chip_root}/src/credentials/examples/ExampleDACs.h",
     "${chip_root}/src/credentials/examples/ExamplePAI.h",
-    "../common/RamStorage.h",
-    "../common/RamStorageKey.h",
   ]
 
   if (chip_with_factory_data == 1) {
