+++ src/platform/nxp/k32w/k32w0/K32W0Config.h
@@ -20,8 +20,7 @@
 
 #include <functional>
 #include <platform/internal/CHIPDeviceLayerInternal.h>
-#include <platform/nxp/k32w/common/RamStorage.h>
-#include <platform/nxp/k32w/common/RamStorageKey.h>
+#include <platform/nxp/k32w/k32w0/RamStorage.h>
 
 namespace chip {
 namespace DeviceLayer {
